Jav/Jav/requirements.txt
scrapy
mysql-connector-python
pymongo
